Implemented ability for admins to add games to the database.
Added two addgame middleware routes in app.js. One is a get request to /addgame linked from the admin nav and accesses the form which adds a game to the database. This form issues a post request to addgame which performs error checking on the data and creates a new game. Also added handlebar views to indicate the success and failure of adding a game.

Co-Authored-By: Christopher Tan <29719007+christan1029@users.noreply.github.com>